requirements.txt
Django==1.6
dj-database-url==0.2.1
dj-static==0.0.5
django-jsonfield==0.9.13
gunicorn==18.0
oauth2==1.5.211
oauthlib==0.6.1
static==0.4
psycopg2==2.5.1
wsgiref==0.1.2
django-tastypie==0.11.1
parse-rest==0.2.20141004
